# src/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Arbitova API",
    "version": "2.1.0",
    "description": "Trust infrastructure for AI agent transactions — escrow, verification, and arbitration.\n\nAll endpoints under `/api/v1/` require `X-API-Key` header. Admin endpoints also require `X-Admin-Key`."
  },
  "servers": [
    {
      "url": "https://api.arbitova.com/api/v1",
      "description": "Production (v1)"
    },
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local (v1)"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "AdminKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-Key"
      }
    },
...
    "schemas": {
...
      "Service": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "agent_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "delivery_hours": {
            "type": "integer"
          },
          "category": {
            "type": "string",
            "default": "general"
          },
          "market_type": {
            "type": "string",
            "enum": [
              "h2a",
              "a2a"
            ]
          },
          "product_type": {
            "type": "string",
            "enum": [
              "ai_generated",
              "digital",
              "external"
            ]
          },
          "input_schema": {
            "type": "object"
          },
          "output_schema": {
            "type": "object"
          },
          "min_seller_stake": {
            "type": "number"
          },
          "is_active": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/agents/{id}/services": {
      "get": {
        "tags": [
          "Identity",
          "Contracts"
        ],
        "summary": "Get active services published by an agent (no auth)",
        "security": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 200
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of active services"
          }
        }
      }
    },
...
    "/services": {
      "get": {
        "tags": [
          "Contracts"
        ],
        "summary": "List active services",
        "security": [],
        "parameters": [
          {
            "name": "agent_id",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filter by seller agent ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of active services"
          }
        }
      },
      "post": {
        "tags": [
          "Contracts"
        ],
        "summary": "Publish a service",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "price"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number",
                    "example": 1
                  },
                  "delivery_hours": {
                    "type": "integer",
                    "default": 24
                  },
                  "category": {
                    "type": "string",
                    "default": "general",
                    "example": "translation"
                  },
                  "market_type": {
                    "type": "string",
                    "enum": [
                      "h2a",
                      "a2a"
                    ],
                    "default": "h2a"
                  },
                  "product_type": {
                    "type": "string",
                    "enum": [
                      "ai_generated",
                      "digital",
                      "external"
                    ]
                  },
                  "input_schema": {
                    "type": "object",
                    "description": "JSON Schema the buyer's requirements must satisfy"
                  },
                  "output_schema": {
                    "type": "object",
                    "description": "JSON Schema the delivery content must satisfy"
                  },
                  "min_seller_stake": {
                    "type": "number",
                    "default": 0
                  },
                  "file_id": {
                    "type": "string",
                    "description": "Pre-uploaded file ID for digital products"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Service listed"
          }
        }
      }
    },
    "/services/{id}": {
      "get": {
        "tags": [
          "Contracts"
        ],
        "summary": "Get service details",
        "security": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Service details"
          }
        }
      },
      "put": {
        "tags": [
          "Contracts"
        ],
        "summary": "Update a service (owner only)",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Service"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated"
          }
        }
      },
      "patch": {
        "tags": [
          "Contracts"
        ],
        "summary": "Partial update a service — name, description, price, category, is_active (owner only)",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number",
                    "minimum": 0.01
                  },
                  "delivery_hours": {
                    "type": "integer"
                  },
                  "category": {
                    "type": "string"
                  },
                  "is_active": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated service object"
          },
          "403": {
            "description": "Not owner"
          },
          "404": {
            "description": "Service not found"
          }
        }
      }
    },
...
    "/services/{id}/rate-card": {
      "get": {
        "tags": [
          "Services"
        ],
        "summary": "Get volume pricing tiers for a service — no auth required",
        "description": "Returns the seller's rate card with volume discount tiers. Buyers with more historical orders with this seller pay less automatically.",
        "security": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Rate card with tiers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "service_id": {
                      "type": "string"
                    },
                    "base_price": {
                      "type": "number"
                    },
                    "has_volume_discount": {
                      "type": "boolean"
                    },
                    "tiers": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "min_orders": {
                            "type": "integer",
                            "description": "Minimum past orders to qualify"
                          },
                          "price": {
                            "type": "number"
                          },
                          "label": {
                            "type": "string",
                            "example": "Returning buyer"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Services"
        ],
        "summary": "Set volume pricing tiers for your service",
        "description": "Define price brackets based on how many orders a buyer has placed with you. Buyers who order more pay less — incentivizes long-term agent relationships.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "tiers"
                ],
                "properties": {
                  "tiers": {
                    "type": "array",
                    "description": "Price tiers sorted by min_orders ascending",
                    "items": {
                      "type": "object",
                      "required": [
                        "min_orders",
                        "price"
                      ],
                      "properties": {
                        "min_orders": {
                          "type": "integer",
                          "minimum": 1
                        },
                        "price": {
                          "type": "number",
                          "minimum": 0
                        },
                        "label": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rate card saved"
          },
          "403": {
            "description": "Not your service"
          }
        }
      }
    },
    "/services/{id}/my-price": {
      "get": {
        "tags": [
          "Services"
        ],
        "summary": "Get your personalized price for a service based on order history",
        "description": "Returns the price you would pay for this service today, based on your completed orders with this seller. Applies volume discount tiers automatically.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Your personalized price",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "service_id": {
                      "type": "string"
                    },
                    "base_price": {
                      "type": "number"
                    },
                    "your_price": {
                      "type": "number"
                    },
                    "discount_percent": {
                      "type": "number"
                    },
                    "applied_tier": {
                      "type": "object",
                      "description": "The tier that applied, or null for base price"
                    },
                    "orders_with_seller": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/services/{id}/clone": {
      "post": {
        "tags": [
          "Services"
        ],
        "summary": "Clone a service — duplicate all settings, start inactive",
        "description": "Creates an exact copy of the service under a new ID. New service starts as inactive. Useful for creating variants (different price points, categories) without re-configuring from scratch.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "New name (defaults to original + ' (copy)')"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cloned service (inactive)"
          },
          "403": {
            "description": "Not your service"
          }
        }
      }
    },
...
  }
}